# server/docs/swagger.yaml
openapi: 3.0.4
info:
  title: Today I Learned v2
  description: Today I Learned is a news-sharing system designed around tags.
  version: "2.0.0"
servers:
  - url: https://api-til.kube.znk.io
...
paths:
...
  /rss:
    get:
      tags:
        - miscellaneous
      summary: Generate RSS feed for user
      description: Generates a Atom 2.0 RSS feed.
      parameters:
        - name: key
          in: query
          description: RSS Feed key
          required: true
          schema:
            type: string
        - name: tags
          in: query
          description: A list of tags, comma-separated
          required: false
          schema:
            type: string
            example: tag1,tag2
      responses:
        '200':
          description: A Atom 2.0 RSS Feed
        '401':
          description: The provided token is invalid.
...
  /posts:
    get:
      tags:
        - posts
      summary: List posts
      description: Get a list of post. A list of post is always paginated.
      security:
        - authorization: []
      parameters:
        - name: tags
          in: query
          description: A list of tags, comma-separated
          required: false
          schema:
            type: string
            example: tag1,tag2
        - name: page
          in: query
          description: The page to retrieve (starts at 0)
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          description: The number of items to retrieve (max is 100)
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: A page of posts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostPaginate'
    post:
      tags:
        - posts
      summary: Create a post
      description: Create a post. The content field is limited to 180 characters.
      security:
        - authorization: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostCreation"
      responses:
        '204':
          description: Created
        '400':
          description: Invalid request.
...
components:
  schemas:
...
    PostPaginate:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Post'
        total_items:
          type: integer
        items_per_page:
          type: integer
        current_page:
          type: integer
        total_pages:
          type: integer
...
    PostCreation:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        link:
          type: string
        tags:
          type: array
          items:
            type: string
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        creation_date:
          type: string
          format: date
        user:
          $ref: '#/components/schemas/UserMinimal'
        title:
          type: string
        content:
          type: string
        link:
          type: string
        tags:
          type: array
          items:
            type: string
...
    UserMinimal:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
  securitySchemes:
    authorization:
      type: apiKey
      name: Authorization
      in: header